+++ function-controller/pkg/controller/autoscaler/simulator/scenarios/combined_scenario.go
@@ -1,44 +1,26 @@
 package scenarios
 
 import (
-	"github.com/projectriff/riff/message-transport/pkg/transport/metrics"
-	"github.com/projectriff/riff/function-controller/pkg/controller/autoscaler/simulator"
 	"math"
 	"time"
+
+	"github.com/projectriff/riff/function-controller/pkg/controller/autoscaler/simulator"
+	"github.com/projectriff/riff/message-transport/pkg/transport/metrics"
 )
 
-const (
-	replicaInitialisationDelaySteps = 15 // 1.5 s delayed initialisation (termination is immediate)
-	containerPullDelaySteps         = 0  // 50 // 5.0 s extra delay in starting the first replica
+type combinedScenario struct{}
 
-	maxWritesPerTick = 40
-)
-
-type CombinedScenario struct {}
-
-func (scenario CombinedScenario) MakeNewSimulation() (metrics.MetricsReceiver, simulator.SimulationUpdater, simulator.ReplicaModel) {
+func MakeNewCombinedScenario() (metrics.MetricsReceiver, simulator.SimulationUpdater, simulator.ReplicaModel) {
 	stubReceiver := newStubReceiver()
 	rm := &replicaModel{initialDelay: containerPullDelaySteps}
+	scenario := &combinedScenario{}
 
-	return stubReceiver, stubReceiver, rm
+	return stubReceiver, scenario, rm
 }
 
-type stubReceiver struct {
-	producerMetricsChan chan metrics.ProducerAggregateMetric
-	consumerMetricsChan chan metrics.ConsumerAggregateMetric
+func (scenario *combinedScenario) UpdateProducerFor(receiver metrics.MetricsReceiver, simulationRound int, queueLen *int64, writes *int) {
+	stubReceiver := receiver.(stubReceiver)
 
-	currentRound int
-}
-
-func (rec *stubReceiver) ProducerMetrics() <-chan metrics.ProducerAggregateMetric {
-	return rec.producerMetricsChan
-}
-
-func (rec *stubReceiver) ConsumerMetrics() <-chan metrics.ConsumerAggregateMetric {
-	return rec.consumerMetricsChan
-}
-
-func (rec *stubReceiver) UpdateProducerFor(simulationRound int, queueLen *int64, writes *int) {
 	numToWrite := 0
 	if simulationRound < 100 {
 		// initial quiet interval
@@ -72,12 +54,14 @@ func (rec *stubReceiver) UpdateProducerFor(simulationRound int, queueLen *int64,
 	*writes = numToWrite
 
 	for i := numToWrite; i > 0; i-- {
-		rec.producerMetricsChan <- metrics.ProducerAggregateMetric{Topic: "topic", Count: 1}
+		stubReceiver.producerMetricsChan <- metrics.ProducerAggregateMetric{Topic: "topic", Count: 1}
 		(*queueLen)++
 	}
 }
 
-func (rec *stubReceiver) UpdatedConsumerFor(simulationRound int, replicas int, queueLen *int64) {
+func (scenario *combinedScenario) UpdatedConsumerFor(receiver metrics.MetricsReceiver, simulationRound int, replicas int, queueLen *int64) {
+	stubReceiver := receiver.(stubReceiver)
+
 	if replicas == 0 {
 		return // nothing doing
 	}
@@ -87,75 +71,7 @@ func (rec *stubReceiver) UpdatedConsumerFor(simulationRound int, replicas int, q
 		numToRead = int(*queueLen)
 	}
 	for i := numToRead; i > 0; i-- {
-		rec.consumerMetricsChan <- metrics.ConsumerAggregateMetric{Topic: "topic", ConsumerGroup: "stub function", Pod: "pod" /*should vary by replica*/ , Count: 1, Interval: time.Millisecond}
+		stubReceiver.consumerMetricsChan <- metrics.ConsumerAggregateMetric{Topic: "topic", ConsumerGroup: "stub function", Pod: "pod" /*should vary by replica*/, Count: 1, Interval: time.Millisecond}
 		(*queueLen)--
 	}
-}
-
-func newStubReceiver() *stubReceiver {
-	return &stubReceiver{
-		producerMetricsChan: make(chan metrics.ProducerAggregateMetric),
-		consumerMetricsChan: make(chan metrics.ConsumerAggregateMetric),
-	}
-}
-
-
-// A replicaModel models the way new replicas take a while to start.
-// An increase of N in the desired number of replicas results in N items being added to `scheduled`. Each item in
-// `scheduled` represents a time (in "ticks") at which the corresponding replica will count towards the actual number of
-// replicas. This isn't quite the same behaviour as k8s, which knows about, and will inform the function controller of,
-// replicas which are still completing their initialisation, but at least it makes the model more realistic than if
-// replicas initialise instantaneously.
-// A decrease in the desired number of replicas is acted upon immediately by removing items from `scheduled` and, if
-// that isn't sufficient, reducing the actual number of replicas.
-type replicaModel struct {
-	currentTime  int // in "ticks"
-	actual       int
-	lastDesired  int
-	scheduled    []int
-	initialDelay int
-}
-
-func (rm *replicaModel) DesireReplicas(desired int) {
-	if desired == rm.lastDesired {
-		return
-	}
-	if desired > rm.lastDesired {
-		// schedule some new replicas with a delay
-		initTime := rm.currentTime + replicaInitialisationDelaySteps + rm.initialDelay
-		rm.initialDelay = 0
-		for i := desired - rm.lastDesired; i > 0; i-- {
-			rm.scheduled = append(rm.scheduled, initTime)
-		}
-	} else {
-		rm.trim(rm.lastDesired - desired)
-	}
-	rm.lastDesired = desired
-}
-
-func (rm *replicaModel) trim(deschedule int) {
-	if deschedule >= len(rm.scheduled) {
-		trimmed := len(rm.scheduled)
-		rm.scheduled = []int{}
-		rm.actual -= deschedule - trimmed
-	} else {
-		rm.scheduled = rm.scheduled[0 : len(rm.scheduled)-deschedule]
-	}
-}
-
-func (rm *replicaModel) ActualReplicas() int {
-	return rm.actual
-}
-
-func (rm *replicaModel) Tick() {
-	rm.currentTime++
-	remaining := []int{}
-	for _, s := range rm.scheduled {
-		if s <= rm.currentTime {
-			rm.actual++
-		} else {
-			remaining = append(remaining, s)
-		}
-	}
-	rm.scheduled = remaining
 }
